refactor add to cart to separate out into multiple components for further clarity
