+++ client/src/components/OverView/SizeDropDown.jsx
@@ -1,26 +1,35 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { CSSTransition, Transition, TransitionGroup } from 'react-transition-group';
+import { DropDownHeader, ListContainer, DropDownItem } from './styles.jsx';
 import styled from 'styled-components';
 
+const SizeDropDownContainer = styled.div`
+  flex-basis: 50%;
+`
 
-const SizeDropDown = () => {
+
+const SizeListContainer = styled(ListContainer)`
+
+`
+
+const SizeDropDown = ({size, sizes, isActive, setIsActive, handleSizeClick}) => {
 
 
   return (
-    <SizeDropDown ref={sizeRef} className="select-size" >
-      <DropdownHeader value={null} onClick={() => {setIsActive(prev => !prev)}}>
+    <SizeDropDownContainer>
+      <DropDownHeader value={null} onClick={() => {setIsActive(prev => !prev)}}>
         {size ? size : 'Select Size'}
-      </DropdownHeader>
+      </DropDownHeader>
         <CSSTransition in={isActive} unmountOnExit timeout={700} classNames='fade'>
 
           <SizeListContainer>
           {sizes.map(size => (
-              <DropdownItem key={size.id} value={size.size} data-id={size.id}  onClick={e => handleSizeClick(e)}>{size.size} </DropdownItem>
+              <DropDownItem key={size.id} value={size.size} data-id={size.id}  onClick={e => handleSizeClick(e)}>{size.size} </DropDownItem>
               ))}
           </SizeListContainer>
 
         </CSSTransition>
-    </SizeDropDown>
+    </SizeDropDownContainer>
   )
 }
